- library changes to build on Win32.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@2646 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/snmplib/vacm.c
@@ -69,7 +69,7 @@ vacm_getViewEntry(const char *viewName,
 	    int mask = 0x80, maskpos = 0;
 	    int oidpos;
             found = 1;
-	    for (oidpos = 0; found && oidpos < vp->viewSubtreeLen; oidpos++) {
+	    for (oidpos = 0; found && oidpos < (int)vp->viewSubtreeLen; oidpos++) {
 		if ((vp->viewMask[maskpos] & mask) != 0) {
 		    if (viewSubtree[oidpos] != vp->viewSubtree[oidpos])
                         found = 0;
